Changes the example file and made gpio.c case insensitive
Added a blink.c file that will blink an led and it will take the name
of the board from the command line so it can run on the dragon, hikey
and bubblegum boards without a recompile.

Also make the gpio.c file case insensitve since it was possible that
the incomming board name could be in odd case since a human could
supply it.